+++ convex/game/inputs.ts
@@ -1,21 +1,13 @@
-import { blocked } from './movement';
-import { characters } from '../../data/characters';
 import { agentInputs } from './agents';
-import { ObjectType, PropertyValidators, v } from 'convex/values';
+import { ObjectType, v } from 'convex/values';
 import { movePlayer } from './movement';
-import { stopConversation } from './conversations';
 import { conversationInputs } from './conversations';
 import { point } from '../util/types';
 import { AiTown } from './aiTown';
+import { inputHandler } from './inputHandler';
+import { joinGame, leaveGame } from './players';
 
-export function inputHandler<ArgsValidator extends PropertyValidators, Return extends any>(def: {
-  args: ArgsValidator;
-  handler: (game: AiTown, now: number, args: ObjectType<ArgsValidator>) => Promise<Return>;
-}) {
-  return def;
-}
 // Join, creating a new player...
-
 export const join = inputHandler({
   args: {
     name: v.string(),
@@ -24,50 +16,14 @@ export const join = inputHandler({
     tokenIdentifier: v.optional(v.string()),
   },
   handler: async (game, now, args) => {
-    const { name, character, description, tokenIdentifier } = args;
-    const players = game.players.allDocuments();
-    let position;
-    for (let attempt = 0; attempt < 10; attempt++) {
-      const candidate = {
-        x: Math.floor(Math.random() * game.map.width),
-        y: Math.floor(Math.random() * game.map.height),
-      };
-      if (blocked(game, now, candidate)) {
-        continue;
-      }
-      position = candidate;
-      break;
-    }
-    if (!position) {
-      throw new Error(`Failed to find a free position!`);
-    }
-    const facingOptions = [
-      { dx: 1, dy: 0 },
-      { dx: -1, dy: 0 },
-      { dx: 0, dy: 1 },
-      { dx: 0, dy: -1 },
-    ];
-    const facing = facingOptions[Math.floor(Math.random() * facingOptions.length)];
-    if (!characters.find((c) => c.name === character)) {
-      throw new Error(`Invalid character: ${character}`);
-    }
-    const locationId = await game.locations.insert(now, {
-      x: position.x,
-      y: position.y,
-      dx: facing.dx,
-      dy: facing.dy,
-      velocity: 0,
-    });
-    const playerId = await game.players.insert({
-      worldId: game.world._id,
-      name,
-      description,
-      active: true,
-      human: tokenIdentifier,
-      character,
-      locationId,
-    });
-    return playerId;
+    return await joinGame(
+      game,
+      now,
+      args.name,
+      args.character,
+      args.description,
+      args.tokenIdentifier,
+    );
   },
 });
 // ...or leave, disabling the specified player.
@@ -76,17 +32,7 @@ const leave = inputHandler({
     playerId: v.id('players'),
   },
   handler: async (game: AiTown, now: number, { playerId }) => {
-    const player = game.players.lookup(playerId);
-    // Stop our conversation if we're leaving the game.
-    const membership = game.conversationMembers.find((m) => m.playerId === playerId);
-    if (membership) {
-      const conversation = game.conversations.find((d) => d._id === membership.conversationId);
-      if (conversation === null) {
-        throw new Error(`Couldn't find conversation: ${membership.conversationId}`);
-      }
-      stopConversation(game, now, conversation);
-    }
-    player.active = false;
+    await leaveGame(game, now, playerId);
     return null;
   },
 });
@@ -107,6 +53,8 @@ export const inputs = {
   moveTo,
   // Inputs for the messaging layer.
   ...conversationInputs,
+  startConversation: conversationInputs.startConversation,
+
   // Inputs for the agent layer.
   ...agentInputs,
 };
