+++ convex/game/players.ts
@@ -4,6 +4,10 @@ import { path, point } from '../util/types';
 import { GameTable } from '../engine/gameTable';
 import { DatabaseWriter } from '../_generated/server';
 import { Doc, Id } from '../_generated/dataModel';
+import { AiTown } from './aiTown';
+import { blocked } from './movement';
+import { characters } from '../../data/characters';
+import { stopConversation } from './conversations';
 
 const pathfinding = v.object({
   destination: point,
@@ -78,4 +82,70 @@ export class Players extends GameTable<'players'> {
   isActive(doc: Doc<'players'>): boolean {
     return doc.active;
   }
+}
+
+export async function joinGame(
+  game: AiTown,
+  now: number,
+  name: string,
+  character: string,
+  description: string,
+  tokenIdentifier?: string,
+) {
+  let position;
+  for (let attempt = 0; attempt < 10; attempt++) {
+    const candidate = {
+      x: Math.floor(Math.random() * game.map.width),
+      y: Math.floor(Math.random() * game.map.height),
+    };
+    if (blocked(game, now, candidate)) {
+      continue;
+    }
+    position = candidate;
+    break;
+  }
+  if (!position) {
+    throw new Error(`Failed to find a free position!`);
+  }
+  const facingOptions = [
+    { dx: 1, dy: 0 },
+    { dx: -1, dy: 0 },
+    { dx: 0, dy: 1 },
+    { dx: 0, dy: -1 },
+  ];
+  const facing = facingOptions[Math.floor(Math.random() * facingOptions.length)];
+  if (!characters.find((c) => c.name === character)) {
+    throw new Error(`Invalid character: ${character}`);
+  }
+  const locationId = await game.locations.insert(now, {
+    x: position.x,
+    y: position.y,
+    dx: facing.dx,
+    dy: facing.dy,
+    velocity: 0,
+  });
+  const playerId = await game.players.insert({
+    worldId: game.world._id,
+    name,
+    description,
+    active: true,
+    human: tokenIdentifier,
+    character,
+    locationId,
+  });
+  return playerId;
+}
+
+export async function leaveGame(game: AiTown, now: number, playerId: Id<'players'>) {
+  const player = game.players.lookup(playerId);
+  // Stop our conversation if we're leaving the game.
+  const membership = game.conversationMembers.find((m) => m.playerId === playerId);
+  if (membership) {
+    const conversation = game.conversations.find((d) => d._id === membership.conversationId);
+    if (conversation === null) {
+      throw new Error(`Couldn't find conversation: ${membership.conversationId}`);
+    }
+    stopConversation(game, now, conversation);
+  }
+  player.active = false;
 }
